Cleaned repositories and added UnitOfWork. Changed price range input in sort menu to decimal. Fixed bug with sorting on some browsers (removed usage of Url.Action). Fixed error handling and changed redirection.

+++ Models/SortMenu.cs
@@ -11,10 +11,10 @@ namespace GoldCap.Models
         public string DescriptionSearch { get; set; }
 
         [Display(Prompt = "From")]
-        public int PriceFrom { get; set; }
+        public decimal PriceFrom { get; set; }
 
         [Display(Prompt = "To")]
-        public int PriceTo { get; set; }
+        public decimal PriceTo { get; set; }
 
         [Range(typeof(DateTime), "1/1/2000", "1/1/2040", ErrorMessage = "Invalid date")]
         public DateTime? DateFrom { get; set; }
